fix examples (ai recomendation)

+++ examples/general/billing.php
@@ -6,7 +6,7 @@ use Mailtrap\MailtrapGeneralClient;
 
 require __DIR__ . '/../vendor/autoload.php';
 
-$accountId = getenv('MAILTRAP_ACCOUNT_ID');
+$accountId = (int) getenv('MAILTRAP_ACCOUNT_ID');
 $config = new Config(getenv('MAILTRAP_API_KEY')); #your API token from here https://mailtrap.io/api-tokens
 $billing = (new MailtrapGeneralClient($config))->billing($accountId);
